feat(aws_savings_plan#Understanding how reserved instances are applied): Understanding how reserved instances are applied
A Reserved Instance that is purchased for a Region is called a regional Reserved Instance, and provides Availability Zone and instance size flexibility.

- The Reserved Instance discount applies to instance usage in any Availability Zone in that Region.
- The Reserved Instance discount applies to instance usage within the instance family, regardless of size—this is known as instance size flexibility.

With instance size flexibility, the Reserved Instance discount applies to instance usage for instances that have the same family, generation, and attribute. The Reserved Instance is applied from the smallest to the largest instance size within the instance family based on the normalization factor.

The discount applies either fully or partially to running instances of the same instance family, depending on the instance size of the reservation, in any Availability Zone in the Region. The only attributes that must be matched are the instance family, tenancy, and platform.

The following table lists the different sizes within an instance family, and the corresponding normalization factor. This scale is used to apply the discounted rate of Reserved Instances to the normalized usage of the instance family.

| Instance size | 	Normalization factor |
| --- | --- |
| nano | 	0.25 |
| micro | 	0.5 |
| small | 	1 |
| medium | 	2 |
| large | 	4 |
| xlarge | 	8 |
| 2xlarge | 	16 |
| 3xlarge | 	24 |
| 4xlarge | 	32 |
| 6xlarge | 	48 |
| 8xlarge | 	64 |
| 9xlarge | 	72 |
| 10xlarge | 	80 |
| 12xlarge | 	96 |
| 16xlarge | 	128 |
| 18xlarge | 	144 |
| 24xlarge | 	192 |
| 32xlarge | 	256 |
| 48xlarge | 	384 |
| 56xlarge | 	448 |
| 112xlarge | 	896 |

For example, a `t2.medium` instance has a normalization factor of `2`. If you purchase a `t2.medium` default tenancy Amazon Linux/Unix Reserved Instance in the US East (N. Virginia) and you have two running `t2.small` instances in your account in that Region, the billing benefit is applied in full to both instances.

Or, if you have one `t2.large` instance running in your account in the US East (N. Virginia) Region, the billing benefit is applied to 50% of the usage of the instance.

Limitations:

- *Supported*: Instance size flexibility is only supported for Regional Reserved Instances.
- *Not supported*: Instance size flexibility is not supported for the following Reserved Instances:
    - Reserved Instances that are purchased for a specific Availability Zone (zonal Reserved Instances)
    - Reserved Instances for G4ad, G4dn, G5, G5g, and Inf1 instances
    - Reserved Instances for Windows Server, Windows Server with SQL Standard, Windows Server with SQL Server Enterprise, Windows Server with SQL Server Web, RHEL, and SUSE Linux Enterprise Server
    - Reserved Instances with dedicated tenancy

feat(gpu): Introduce GPU

[GPU](https://en.wikipedia.org/wiki/Graphics_processing_unit) or Graphic Processing Unit is a specialized electronic circuit initially designed to accelerate computer graphics and image processing (either on a video card or embedded on motherboards, mobile phones, personal computers, workstations, and game consoles).

For years I've wanted to buy a graphic card but I've been stuck in the problem that I don't have a desktop. I have a X280 lenovo laptop used to work and personal use with an integrated card that has let me so far to play old games such as [King Arthur Gold](kag.md) or [Age of Empires II](age_of_empires.md), but has hard times playing "newer" games such as It takes two. Last year I also bought a [NAS](nas.md) with awesome hardware. So it makes no sense to buy a desktop just for playing.

Now that I host [Jellyfin](jellyfin.md) on the NAS and that machine learning is on the hype with a lot of interesting solutions that can be self-hosted (whisper, chatgpt similar solutions...), it starts to make sense to add a GPU to the server. What made me give the step is that you can also self-host a gaming server to stream to any device! It makes so much sense to have all the big guns inside the NAS and stream the content to the less powerful devices.

That way if you host services, you make the most use of the hardware.

feat(grapheneos#Disable Bluetooth and Wifi once it's disconnected): Disable Bluetooth and Wifi once it's disconnected

If you don't want to go spreading your SSIDs you can configure graphene to disable wifi and bluetooth X minutes after loosing connection.

For Wifi:

- Go to Settings > network & internet > internet > network preferences
- Select Turn off Wi-Fi automatically, for example after 2 minutes

For Bluetooth:

- Go to Settings > connected devices > connection preferences>  bluetooth
- Select Turn Bluetooth off automatically, for example after 2 minutes

feat(moonlight): Introduce moonlight

[Moonlight](https://github.com/moonlight-stream/moonlight-docs/wiki) is an open source client implementation of NVIDIA GameStream that allows you to to stream your collection of games and apps from your GameStream-compatible PC to another device on your network or the Internet. You can play your favorite games on your PC, phone, tablet, or TV with Moonlight..

References:

- [Home](https://moonlight-stream.org/)
- [Docs](https://github.com/moonlight-stream/moonlight-docs/wiki)

feat(sqlite#Order by a column descending): Order by a column descending

```sql
SELECT
   select_list
FROM
   table
ORDER BY
    column_1 ASC,
    column_2 DESC;
```

feat(promtail): Introduce Promtail

[Promtail](https://grafana.com/docs/loki/latest/send-data/promtail/) is an agent which ships the contents of local logs to a [Loki](loki.md) instance.

It is usually deployed to every machine that runs applications which need to be monitored.

It primarily:

- Discovers targets
- Attaches labels to log streams
- Pushes them to the Loki instance.